Factor out a trace() decorator factory.
